moving mielens calculation functions to separate file
