+++ holopy/scattering/tests/test_mielens_calculations.py
@@ -2,8 +2,9 @@ import unittest
 
 import numpy as np
 from scipy.special import jn_zeros
+from nose.plugins.attrib import attr
 
-from ..theory import mielens
+from ..theory import mielensfunctions
 
 TOLS = {'atol': 1e-10, 'rtol': 1e-10}
 MEDTOLS = {"atol": 1e-6, "rtol": 1e-6}
@@ -11,12 +12,14 @@ SOFTTOLS = {'atol': 1e-3, 'rtol': 1e-3}
 
 
 class TestMieLensCalculator(unittest.TestCase):
+    @attr("fast")
     def test_fields_nonzero(self):
         field1_x, field1_y = evaluate_scattered_field_in_lens()
         should_be_nonzero = [np.linalg.norm(f) for f in [field1_x, field1_y]]
         should_be_false = [np.isclose(v, 0, **TOLS) for v in should_be_nonzero]
         self.assertFalse(any(should_be_false))
 
+    @attr("fast")
     def test_scatteredfield_linear_at_low_contrast(self):
         dn1 = 1e-3
         dn2 = 5e-4
@@ -37,9 +40,10 @@ class TestMieLensCalculator(unittest.TestCase):
 
 
 class TestFarfieldMieInterpolator(unittest.TestCase):
+    @attr("fast")
     def test_interpolator_1_accuracy(self):
         theta = np.linspace(0, 1.5, 1000)
-        interpolator = mielens.FarfieldMieInterpolator(i=1)
+        interpolator = mielensfunctions.FarfieldMieInterpolator(i=1)
 
         exact = interpolator._eval(theta)
         approx = interpolator(theta)
@@ -48,9 +52,10 @@ class TestFarfieldMieInterpolator(unittest.TestCase):
         is_ok = np.allclose(exact / rescale, approx / rescale, **MEDTOLS)
         self.assertTrue(is_ok)
 
+    @attr("fast")
     def test_interpolator_2_accuracy(self):
         theta = np.linspace(0, 1.5, 1000)
-        interpolator = mielens.FarfieldMieInterpolator(i=2)
+        interpolator = mielensfunctions.FarfieldMieInterpolator(i=2)
 
         exact = interpolator._eval(theta)
         approx = interpolator(theta)
@@ -59,12 +64,13 @@ class TestFarfieldMieInterpolator(unittest.TestCase):
         is_ok = np.allclose(exact / rescale, approx / rescale, **MEDTOLS)
         self.assertTrue(is_ok)
 
+    @attr("fast")
     def test_interpolator_maxl_accuracy(self):
         theta = np.linspace(0, 1.5, 1000)
-        interpolator_low_l = mielens.FarfieldMieInterpolator(i=1)
+        interpolator_low_l = mielensfunctions.FarfieldMieInterpolator(i=1)
 
         higher_l = np.ceil(interpolator_low_l.size_parameter * 8).astype('int')
-        interpolator_higher_l = mielens.FarfieldMieInterpolator(
+        interpolator_higher_l = mielensfunctions.FarfieldMieInterpolator(
             i=1, max_l=higher_l)
 
         exact = interpolator_low_l._eval(theta)
@@ -76,37 +82,42 @@ class TestFarfieldMieInterpolator(unittest.TestCase):
 
 
 class TestGaussQuad(unittest.TestCase):
+    @attr("fast")
     def test_constant_integrand(self):
         f = lambda x: np.ones_like(x, dtype='float')
         should_be_one = integrate_like_mielens(f, [3, 4])
         self.assertTrue(np.isclose(should_be_one, 1.0, **TOLS))
 
+    @attr("fast")
     def test_linear_integrand(self):
         f = lambda x: x
         should_be_onehalf = integrate_like_mielens(f, [0, 1])
         self.assertTrue(np.isclose(should_be_onehalf, 0.5, **TOLS))
 
+    @attr("fast")
     def test_transcendental_integrand(self):
         should_be_two = integrate_like_mielens(np.sin, [0, np.pi])
         self.assertTrue(np.isclose(should_be_two, 2.0, **TOLS))
 
 
 class TestJ2(unittest.TestCase):
+    @attr("fast")
     def test_fastj2_zeros(self):
         j2_zeros = jn_zeros(2, 50)
-        should_be_zero = mielens.j2(j2_zeros)
+        should_be_zero = mielensfunctions.j2(j2_zeros)
         self.assertTrue(np.allclose(should_be_zero, 0, atol=1e-13))
 
+    @attr("fast")
     def test_fastj2_notzero(self):
         # We pick some values that should not be zero or close to it
         j0_zeros = jn_zeros(0, 50)  # some conjecture -- bessel functions
         # only share the zero at z=0, which is not for j0
-        should_not_be_zero = mielens.j2(j0_zeros)
+        should_not_be_zero = mielensfunctions.j2(j0_zeros)
         self.assertFalse(np.isclose(should_not_be_zero, 0, atol=1e-10).any())
 
 
 def evaluate_scattered_field_in_lens(delta_index=0.1, size_parameter=0.1):
-    miecalculator = mielens.MieLensCalculator(
+    miecalculator = mielensfunctions.MieLensCalculator(
         particle_kz=10, index_ratio=1.0 + delta_index,
         size_parameter=size_parameter, lens_angle=0.9)
     krho = np.linspace(0, 30, 300)
@@ -116,7 +127,7 @@ def evaluate_scattered_field_in_lens(delta_index=0.1, size_parameter=0.1):
 
 
 def integrate_like_mielens(function, bounds):
-    pts, wts = mielens.gauss_legendre_pts_wts(bounds[0], bounds[1])
+    pts, wts = mielensfunctions.gauss_legendre_pts_wts(bounds[0], bounds[1])
     return (function(pts) * wts).sum()
 
 if __name__ == '__main__':
